requirements.txt
anthropic>=0.40.0
tweepy>=4.14.0
feedparser>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
plotly>=5.18.0
kaleido>=0.2.1
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
apscheduler>=3.10.0
itsdangerous>=2.1.0
httpx>=0.27.0
